tRCP: move client code to src/common/utils
