+++ src/modules/llms/vendors/azure/AzureSourceSetup.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import { Alert, Box, Button, FormControl, FormLabel, Input, Typography } from '@mui/joy';
 import SyncIcon from '@mui/icons-material/Sync';
 
-import { apiQuery } from '~/modules/trpc/trpc.client';
+import { apiQuery } from '~/common/util/trpc.client';
 
 import { FormInputKey } from '~/common/components/FormInputKey';
 import { Link } from '~/common/components/Link';
