+++ src/modules/llms/vendors/localai/LocalAISourceSetup.tsx
@@ -5,7 +5,7 @@ import { Box, Button, Typography } from '@mui/joy';
 import FileDownloadIcon from '@mui/icons-material/FileDownload';
 import SyncIcon from '@mui/icons-material/Sync';
 
-import { apiQuery } from '~/modules/trpc/trpc.client';
+import { apiQuery } from '~/common/util/trpc.client';
 
 import { FormInputKey } from '~/common/components/FormInputKey';
 import { InlineError } from '~/common/components/InlineError';
